# openapi.json
{"openapi":"3.1.0","info":{"title":"Dynamic LangGraph Agent API","description":"API for building and executing dynamic LangGraph agents","version":"2.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/finetune/jobs":{"post":{"tags":["finetune"],"summary":"Create Finetune Job","operationId":"create_finetune_job_finetune_jobs_post","parameters":[{"name":"model_name","in":"query","required":true,"schema":{"type":"string","title":"Model Name"}},{"name":"dataset_path","in":"query","required":true,"schema":{"type":"string","title":"Dataset Path"}},{"name":"epochs","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"Epochs"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["finetune"],"summary":"List Jobs","operationId":"list_jobs_finetune_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/finetune/jobs/{job_id}":{"get":{"tags":["finetune"],"summary":"Get Job Status","operationId":"get_job_status_finetune_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/upload":{"post":{"tags":["rag"],"summary":"Upload Document","operationId":"upload_document_rag_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_rag_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/query":{"post":{"tags":["rag"],"summary":"Query Knowledge Base","operationId":"query_knowledge_base_rag_query_post","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"n_results","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"N Results"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/servers":{"get":{"tags":["mcp"],"summary":"List Servers","operationId":"list_servers_mcp_servers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["mcp"],"summary":"Connect Server","operationId":"connect_server_mcp_servers_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MCPServerConfig"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools":{"get":{"tags":["mcp"],"summary":"List Tools","operationId":"list_tools_mcp_tools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mcp/tools/call":{"post":{"tags":["mcp"],"summary":"Call Tool","operationId":"call_tool_mcp_tools_call_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolCallRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graphs":{"post":{"summary":"Create Graph","description":"Creates or updates a graph configuration and compiles it.","operationId":"create_graph_graphs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GraphConfig"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graphs/{graph_id}":{"get":{"summary":"Get Graph","description":"Retrieves a graph configuration.","operationId":"get_graph_graphs__graph_id__get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graphs/{graph_id}/execute":{"post":{"summary":"Execute Graph","description":"Executes a graph synchronously (for MVP).","operationId":"execute_graph_graphs__graph_id__execute_post","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/nodes/generate":{"post":{"summary":"Generate Node Code","description":"Generates Python code for a node based on natural language description.","operationId":"generate_node_code_nodes_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateNodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_document_rag_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_rag_upload_post"},"DocumentResponse":{"properties":{"filename":{"type":"string","title":"Filename"},"status":{"type":"string","title":"Status"},"chunks_added":{"type":"integer","title":"Chunks Added"}},"type":"object","required":["filename","status","chunks_added"],"title":"DocumentResponse"},"EdgeConfig":{"properties":{"source":{"type":"string","title":"Source","description":"Source node ID"},"target":{"type":"string","title":"Target","description":"Target node ID"},"condition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Condition","description":"Condition for conditional edges (optional)"}},"type":"object","required":["source","target"],"title":"EdgeConfig","description":"Configuration for a directed edge between nodes."},"ExecuteRequest":{"properties":{"input":{"additionalProperties":true,"type":"object","title":"Input"},"config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Config"}},"type":"object","required":["input"],"title":"ExecuteRequest"},"GenerateNodeRequest":{"properties":{"description":{"type":"string","title":"Description"}},"type":"object","required":["description"],"title":"GenerateNodeRequest"},"GraphConfig":{"properties":{"id":{"type":"string","title":"Id","description":"Unique identifier for the graph"},"name":{"type":"string","title":"Name","description":"Human-readable name of the graph"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Description of the graph"},"nodes":{"items":{"$ref":"#/components/schemas/NodeConfig"},"type":"array","title":"Nodes","description":"List of nodes"},"edges":{"items":{"$ref":"#/components/schemas/EdgeConfig"},"type":"array","title":"Edges","description":"List of edges"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata","description":"Additional metadata"}},"type":"object","required":["id","name","nodes","edges"],"title":"GraphConfig","description":"Complete configuration for a dynamic LangGraph."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MCPServerConfig":{"properties":{"name":{"type":"string","title":"Name"},"command":{"type":"string","title":"Command"},"args":{"items":{"type":"string"},"type":"array","title":"Args","default":[]},"env":{"additionalProperties":{"type":"string"},"type":"object","title":"Env","default":{}}},"type":"object","required":["name","command"],"title":"MCPServerConfig"},"NodeConfig":{"properties":{"id":{"type":"string","title":"Id","description":"Unique identifier for the node"},"type":{"type":"string","enum":["LLMNode","CodeNode","RAGNode","HumanNode"],"title":"Type","description":"Type of the node"},"config":{"additionalProperties":true,"type":"object","title":"Config","description":"Specific configuration for the node type"},"position":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Position","description":"UI position (x, y)"}},"type":"object","required":["id","type"],"title":"NodeConfig","description":"Configuration for a single node in the graph."},"ToolCallRequest":{"properties":{"server_name":{"type":"string","title":"Server Name"},"tool_name":{"type":"string","title":"Tool Name"},"arguments":{"additionalProperties":true,"type":"object","title":"Arguments","default":{}}},"type":"object","required":["server_name","tool_name"],"title":"ToolCallRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}